+++ utils/is-group-member.js
@@ -3,13 +3,7 @@
 const os = require('os');
 
 const posixCmd = user => (['groups', [user]]);
-const win32Cmd = user => ([
-  'powershell.exe',
-  [
-    '-Command',
-    `Get-LocalGroup | Where-Object { ($_ | Get-LocalGroupMember | Where-Object { $_.Name -like "*\\${user}" }).Count -gt 0 } | Select-Object -Property Name`, // eslint-disable-line max-len
-  ],
-]);
+const win32Cmd = () => (['powershell.exe', ['-Command', 'whoami /GROUPS /FO CSV']]);
 
 module.exports = (group, user, platform = process.platform) => {
   // @TODO: throw error if no group specified?
@@ -17,7 +11,8 @@ module.exports = (group, user, platform = process.platform) => {
   if (!user) user = os.userInfo().username;
 
   // get the result of the membership command
-  const cmd = platform === 'win32' ? win32Cmd(user) : posixCmd(user);
+  // @TODO: support user on win32
+  const cmd = platform === 'win32' ? win32Cmd() : posixCmd(user);
   const {status, stdout, stderr} = require('./spawn-sync-stringer')(...cmd);
 
   // if we failed for some reason
@@ -31,12 +26,17 @@ module.exports = (group, user, platform = process.platform) => {
 
   // if windows we have a long command to check
   if (platform === 'win32') {
-    const groups = stdout
+    const ids = stdout
       .split(os.EOL)
-      .map(group => group.trim())
-      .filter(group => group !== 'Name' && group !== '----')
-      .map(group => group.toUpperCase());
-    return groups.includes(group.toUpperCase());
+      .map(line => line.split(',')
+      .map(item => item.replace(/^['"]|['"]$/g, ''))
+      .filter((item, index) => index === 0 || item.startsWith('S-')))
+      .flat()
+      .map(id => id.toUpperCase());
+
+    const matches = ids.filter(id => id === group.toUpperCase() || id.endsWith(group.toUpperCase()));
+
+    return matches.length > 0;
   }
 
   // otherwise false?